Divide uml parser class into smaller classes and enable object diagram
