+++ src/main/java/de/tum/in/www1/artemis/service/compass/umlmodel/object/UMLObject.java
@@ -5,8 +5,6 @@ import java.util.List;
 import de.tum.in.www1.artemis.service.compass.strategy.NameSimilarity;
 import de.tum.in.www1.artemis.service.compass.umlmodel.Similarity;
 import de.tum.in.www1.artemis.service.compass.umlmodel.UMLElement;
-import de.tum.in.www1.artemis.service.compass.umlmodel.classdiagram.UMLAttribute;
-import de.tum.in.www1.artemis.service.compass.umlmodel.classdiagram.UMLMethod;
 
 public class UMLObject extends UMLElement {
 
@@ -14,11 +12,11 @@ public class UMLObject extends UMLElement {
 
     private final String name;
 
-    private final List<UMLAttribute> attributes;
+    private final List<UMLObjectAttribute> attributes;
 
-    private final List<UMLMethod> methods;
+    private final List<UMLObjectMethod> methods;
 
-    public UMLObject(String name, List<UMLAttribute> attributes, List<UMLMethod> methods, String jsonElementID) {
+    public UMLObject(String name, List<UMLObjectAttribute> attributes, List<UMLObjectMethod> methods, String jsonElementID) {
         super(jsonElementID);
         this.name = name;
         this.attributes = attributes;
@@ -31,28 +29,13 @@ public class UMLObject extends UMLElement {
      * Sets the parent of all child elements. The child elements of a UML object are its attributes and methods.
      */
     private void setParentForChildElements() {
-        for (UMLAttribute attribute : attributes) {
+        for (UMLObjectAttribute attribute : attributes) {
             attribute.setParentElement(this);
         }
 
-        for (UMLMethod method : methods) {
+        for (UMLObjectMethod method : methods) {
             method.setParentElement(this);
         }
-    }
-
-    @Override
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public String getType() {
-        return UML_OBJECT_TYPE;
-    }
-
-    @Override
-    public String toString() {
-        return "Object " + name;
     }
 
     @Override
@@ -101,18 +84,99 @@ public class UMLObject extends UMLElement {
         similarity += weight * similarity(referenceObject);
 
         // check attributes
-        for (UMLAttribute attribute : attributes) {
+        for (UMLObjectAttribute attribute : attributes) {
             double similarityValue = referenceObject.similarAttributeScore(attribute);
             similarity += weight * similarityValue;
         }
 
         // check methods
-        for (UMLMethod method : methods) {
+        for (UMLObjectMethod method : methods) {
             double similarityValue = referenceObject.similarMethodScore(method);
             similarity += weight * similarityValue;
         }
 
         return ensureSimilarityRange(similarity);
+    }
+
+    /**
+     * Returns the maximum similarity score of the given attribute and the list of attributes of this class, i.e. the similarity between the reference attribute and the most
+     * similar attribute of this class.
+     *
+     * @param referenceAttribute the reference attribute that should be compared to the attributes of this class
+     * @return the maximum similarity score of the reference attribute and the list of attributes of this class
+     */
+    private double similarAttributeScore(UMLObjectAttribute referenceAttribute) {
+        return similarScore(referenceAttribute, attributes);
+    }
+
+    /**
+     * Returns the maximum similarity score of the given method and the list of methods of this class, i.e. the similarity between the reference method and the most
+     * similar method of this class.
+     *
+     * @param referenceMethod the reference method that should be compared to the methods of this class
+     * @return the maximum similarity score of the reference method and the list of method of this class
+     */
+    private double similarMethodScore(UMLObjectMethod referenceMethod) {
+        return similarScore(referenceMethod, methods);
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public String getType() {
+        return UML_OBJECT_TYPE;
+    }
+
+    @Override
+    public String toString() {
+        return "Object " + name;
+    }
+
+    /**
+     * Checks if the UML class or one of its methods/attributes has the given element ID and returns the corresponding element. Otherwise, it returns null.
+     *
+     * @param jsonElementId the id of the UML element that should be returned
+     * @return the UML element if one could be found for the given id, null otherwise
+     */
+    UMLElement getElementByJSONID(String jsonElementId) {
+        if (getJSONElementID().equals(jsonElementId)) {
+            return this;
+        }
+
+        for (UMLObjectAttribute umlAttribute : getAttributes()) {
+            if (umlAttribute.getJSONElementID().equals(jsonElementId)) {
+                return umlAttribute;
+            }
+        }
+
+        for (UMLObjectMethod umlMethod : getMethods()) {
+            if (umlMethod.getJSONElementID().equals(jsonElementId)) {
+                return umlMethod;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Get the list of UML attributes of this object.
+     *
+     * @return the list of attributes
+     */
+    public List<UMLObjectAttribute> getAttributes() {
+        return attributes;
+    }
+
+    /**
+     * Get the list of UML methods of this object.
+     *
+     * @return the list of methods
+     */
+    public List<UMLObjectMethod> getMethods() {
+        return methods;
     }
 
     /**
@@ -124,25 +188,4 @@ public class UMLObject extends UMLElement {
         return attributes.size() + methods.size() + 1;
     }
 
-    /**
-     * Returns the maximum similarity score of the given attribute and the list of attributes of this class, i.e. the similarity between the reference attribute and the most
-     * similar attribute of this class.
-     *
-     * @param referenceAttribute the reference attribute that should be compared to the attributes of this class
-     * @return the maximum similarity score of the reference attribute and the list of attributes of this class
-     */
-    private double similarAttributeScore(UMLAttribute referenceAttribute) {
-        return similarScore(referenceAttribute, attributes);
-    }
-
-    /**
-     * Returns the maximum similarity score of the given method and the list of methods of this class, i.e. the similarity between the reference method and the most
-     * similar method of this class.
-     *
-     * @param referenceMethod the reference method that should be compared to the methods of this class
-     * @return the maximum similarity score of the reference method and the list of method of this class
-     */
-    private double similarMethodScore(UMLMethod referenceMethod) {
-        return similarScore(referenceMethod, methods);
-    }
 }
