+++ src/main/java/de/tum/in/www1/artemis/service/compass/umlmodel/parsers/ComponentDiagramParser.java
@@ -1,0 +1,124 @@
+package de.tum.in.www1.artemis.service.compass.umlmodel.parsers;
+
+import static de.tum.in.www1.artemis.service.compass.utils.JSONMapping.*;
+
+import java.io.IOException;
+import java.util.*;
+
+import org.apache.commons.lang3.EnumUtils;
+
+import com.google.common.base.CaseFormat;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+
+import de.tum.in.www1.artemis.service.compass.umlmodel.UMLElement;
+import de.tum.in.www1.artemis.service.compass.umlmodel.component.UMLComponent;
+import de.tum.in.www1.artemis.service.compass.umlmodel.component.UMLComponentDiagram;
+import de.tum.in.www1.artemis.service.compass.umlmodel.component.UMLComponentInterface;
+import de.tum.in.www1.artemis.service.compass.umlmodel.component.UMLComponentRelationship;
+
+public class ComponentDiagramParser {
+
+    /**
+     * Create a UML component diagram from the model and relationship elements given as JSON arrays. It parses the JSON objects to corresponding Java objects and creates a
+     * component diagram containing these UML model elements.
+     *
+     * @param modelElements the model elements as JSON array
+     * @param relationships the relationship elements as JSON array
+     * @param modelSubmissionId the ID of the corresponding modeling submission
+     * @return a UML component diagram containing the parsed model elements and relationships
+     * @throws IOException when no corresponding model elements could be found for the source and target IDs in the relationship JSON objects
+     */
+    protected static UMLComponentDiagram buildComponentDiagramFromJSON(JsonArray modelElements, JsonArray relationships, long modelSubmissionId) throws IOException {
+        Map<String, UMLComponent> umlComponentMap = new HashMap<>();
+        Map<String, UMLComponentInterface> umlComponentInterfaceMap = new HashMap<>();
+        Map<String, UMLElement> allUmlElementsMap = new HashMap<>();
+        List<UMLComponentRelationship> umlComponentRelationshipList = new ArrayList<>();
+
+        // owners might not yet be available, therefore we need to store them in a map first before we can resolve them
+        Map<UMLElement, String> ownerRelationships = new HashMap<>();
+
+        // loop over all JSON elements and create the UML objects
+        for (JsonElement jsonElement : modelElements) {
+            JsonObject jsonObject = jsonElement.getAsJsonObject();
+            UMLElement umlElement = null;
+            String elementType = jsonObject.get(ELEMENT_TYPE).getAsString();
+            if (UMLComponent.UML_COMPONENT_TYPE.equals(elementType)) {
+                UMLComponent umlComponent = parseComponent(jsonObject);
+                umlComponentMap.put(umlComponent.getJSONElementID(), umlComponent);
+                umlElement = umlComponent;
+            }
+            else if (UMLComponentInterface.UML_COMPONENT_INTERFACE_TYPE.equals(elementType)) {
+                UMLComponentInterface umlComponentInterface = parseComponentInterface(jsonObject);
+                umlComponentInterfaceMap.put(umlComponentInterface.getJSONElementID(), umlComponentInterface);
+                umlElement = umlComponentInterface;
+            }
+            if (umlElement != null) {
+                allUmlElementsMap.put(umlElement.getJSONElementID(), umlElement);
+                findOwner(ownerRelationships, jsonObject, umlElement);
+            }
+        }
+
+        // now we can resolve the owners: for this diagram type, only uml components can be the actual owner
+        resolveParentComponent(allUmlElementsMap, ownerRelationships);
+
+        // loop over all JSON control flow elements and create UML communication links
+        for (JsonElement rel : relationships) {
+            Optional<UMLComponentRelationship> componentRelationship = parseComponentRelationship(rel.getAsJsonObject(), allUmlElementsMap);
+            componentRelationship.ifPresent(umlComponentRelationshipList::add);
+        }
+
+        return new UMLComponentDiagram(modelSubmissionId, new ArrayList<>(umlComponentMap.values()), new ArrayList<>(umlComponentInterfaceMap.values()),
+                umlComponentRelationshipList);
+    }
+
+    protected static Optional<UMLComponentRelationship> parseComponentRelationship(JsonObject relationshipJson, Map<String, UMLElement> allUmlElementsMap) throws IOException {
+
+        String relationshipType = relationshipJson.get(RELATIONSHIP_TYPE).getAsString();
+        relationshipType = CaseFormat.UPPER_CAMEL.to(CaseFormat.UPPER_UNDERSCORE, relationshipType);
+
+        if (!EnumUtils.isValidEnum(UMLComponentRelationship.UMLComponentRelationshipType.class, relationshipType)) {
+            return Optional.empty();
+        }
+
+        UMLElement source = UMLModelParser.findElement(relationshipJson, allUmlElementsMap, RELATIONSHIP_SOURCE);
+        UMLElement target = UMLModelParser.findElement(relationshipJson, allUmlElementsMap, RELATIONSHIP_TARGET);
+
+        if (source != null && target != null) {
+            UMLComponentRelationship newComponentRelationship = new UMLComponentRelationship(source, target,
+                    UMLComponentRelationship.UMLComponentRelationshipType.valueOf(relationshipType), relationshipJson.get(ELEMENT_ID).getAsString());
+            return Optional.of(newComponentRelationship);
+        }
+        else {
+            throw new IOException("Relationship source or target not part of model!");
+        }
+    }
+
+    protected static UMLComponentInterface parseComponentInterface(JsonObject componentInterfaceJson) {
+        String componentInterfaceName = componentInterfaceJson.get(ELEMENT_NAME).getAsString();
+        return new UMLComponentInterface(componentInterfaceName, componentInterfaceJson.get(ELEMENT_ID).getAsString());
+    }
+
+    protected static UMLComponent parseComponent(JsonObject componentJson) {
+        String componentName = componentJson.get(ELEMENT_NAME).getAsString();
+        return new UMLComponent(componentName, componentJson.get(ELEMENT_ID).getAsString());
+    }
+
+    protected static void resolveParentComponent(Map<String, UMLElement> allUmlElementsMap, Map<UMLElement, String> ownerRelationships) {
+        for (var ownerEntry : ownerRelationships.entrySet()) {
+            String ownerId = ownerEntry.getValue();
+            UMLElement umlElement = ownerEntry.getKey();
+            UMLElement parentComponent = allUmlElementsMap.get(ownerId);
+            umlElement.setParentElement(parentComponent);
+        }
+    }
+
+    protected static void findOwner(Map<UMLElement, String> ownerRelationships, JsonObject jsonObject, UMLElement umlElement) {
+        if (jsonObject.has(ELEMENT_OWNER) && !jsonObject.get(ELEMENT_OWNER).isJsonNull()) {
+            String ownerId = jsonObject.get(ELEMENT_OWNER).getAsString();
+            ownerRelationships.put(umlElement, ownerId);
+        }
+    }
+
+}
