+++ src/main/webapp/app/exercises/modeling/manage/modeling-exercise-update.component.html
@@ -135,14 +135,7 @@
                 <option value="Flowchart">{{ 'artemisApp.DiagramType.Flowchart' | artemisTranslate }}</option>
             </select>
         </div>
-        <div
-            class="form-group"
-            *ngIf="
-                semiAutomaticAssessmentAvailable &&
-                !isExamMode &&
-                (this.modelingExercise.diagramType === UMLDiagramType.ClassDiagram || this.modelingExercise.diagramType === UMLDiagramType.ActivityDiagram)
-            "
-        >
+        <div class="form-group" *ngIf="semiAutomaticAssessmentAvailable && !isExamMode">
             <div class="form-check custom-control custom-checkbox">
                 <input
                     type="checkbox"
